moving tree classes to ResultTree.cs
